refactor(operations): clean up operations managing code
- removed unused modules and imports
- simplified function signatures by removing unnecessary generic parameters
- consolidated template operations into the TemplateOperator
- added helper methods for fetching previous template versions
- improved logging for better traceability
- removed redundant code and improved readability

the changes aim to enhance code maintainability and readability by reducing
complexity and eliminating unused code.
